requirements.txt
openai
Pillow
streamlit
streamlit_carousel